fix: moved catalog product list to separate component
