--- src/app/[category]/components/ProductList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProductList';
