+++ src/app/[category]/page.tsx
@@ -4,10 +4,10 @@ import Toolbar from '@/modules/Toolbar';
 import { getAllCategories } from '@/shared/api/categoryApi';
 
 import Overview from '@/components/Overview';
-import ProductList from './components/ProductList';
 import { notFound } from 'next/navigation';
 import Breadcrumbs from '@/components/Breadcrumbs';
 import { ToolbarProvider } from '@/modules/Toolbar/ToolbarContext';
+import ProductList from '@/modules/CatalogProductList';
 
 export async function generateStaticParams() {
   const { content: categories } = await getAllCategories();
