--- src/app/products/components/ProductList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProductList';
